+++ lib/togls/errors.rb
@@ -6,4 +6,5 @@ module Togls
   class NotImplemented < Error; end
   class FeatureAlreadyDefined < Error; end
   class RuleTypeAlreadyDefined < Error; end
+  class RuleFeatureTargetTypeMismatch < Error; end
 end