+++ lib/togls/toggle.rb
@@ -5,8 +5,7 @@ module Togls
   # responsibility is binding a specific rule to a specific feature. Toggle's by
   # default are associated with a boolean rule initialized to false.
   class Toggle
-    attr_reader :feature
-    attr_accessor :rule
+    attr_reader :feature, :rule
 
     def initialize(feature)
       @feature = feature
@@ -15,6 +14,23 @@ module Togls
 
     def id
       @feature.id
+    end
+
+    def rule=(rule)
+      raise Togls::RuleFeatureTargetTypeMismatch unless target_matches?(rule)
+      @rule = rule
+    end
+
+    # feature target type | rule target type | match?
+    # ------------------------------------------------
+    # any (aka none)      | any              | true
+    # something (foo)     | any              | true
+    # any (aka none)      | something (foo)  | false
+    # something (foo)     | something (foo)  | true
+    # something (foo)     | something (bar)  | false
+    def target_matches?(rule)
+      @feature.target_type == rule.class.target_type ||
+        rule.class.target_type == Togls::TargetTypes::ANY
     end
 
     def on?(target = nil)